+++ psutil/_psosx.py
@@ -67,6 +67,7 @@ pmmap_ext = namedtuple(
 
 # set later from __init__.py
 NoSuchProcess = None
+ZombieProcess = None
 AccessDenied = None
 TimeoutExpired = None
 
@@ -197,10 +198,14 @@ def wrap_exceptions(fun):
             return fun(self, *args, **kwargs)
         except OSError as err:
             # support for private module import
-            if NoSuchProcess is None or AccessDenied is None:
+            if (NoSuchProcess is None or AccessDenied is None or
+                    ZombieProcess is None):
                 raise
             if err.errno == errno.ESRCH:
-                raise NoSuchProcess(self.pid, self._name)
+                if not pid_exists(self.pid):
+                    raise NoSuchProcess(self.pid, self._name)
+                else:
+                    raise ZombieProcess(self.pid, self._name)
             if err.errno in (errno.EPERM, errno.EACCES):
                 raise AccessDenied(self.pid, self._name)
             raise