+++ HISTORY.rst
@@ -18,6 +18,8 @@ Bug tracker at https://github.com/giampaolo/psutil/issues
 
 **Bug fixes**
 
+- #428: [all UNIXes except Linux] correct handling of zombie processes;
+  introduced new ZombieProcess exception class.
 - #555: [Linux] psutil.users() correctly handles ":0" as an alias for
   "localhost"
 - #579: [Windows] Fixed open_files() for PID>64K.